backend/process-pdf/requirements.txt
boto3>=1.26.0
psycopg2-binary>=2.9.0
PyPDF2>=3.0.0
